backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.27.0
python-dotenv==1.0.0
letta-client==0.1.0
anthropic==0.7.8
chromadb==0.4.22
